jwt-convenevc-com/jwt_backend_py/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pkg_resources==0.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
zipp==3.16.2
